feat(signup.js): 회원가입 페이지 이벤트 핸들러 제작
1. TEST_USER 이메일로 가입하려는 경우 "이미 사용 중인 이메일입니다." 출력
2. 두 비밀번호 인풋에 입력한 값이 다를 경우 "비밀번호가 일치하지 않아요" 출력
3. 세 인풋 모두에서 에러메세지가 출력되지 않은 경우 회원가입 성공 후 '/folder'로 이동
