+++ signup.html
@@ -27,7 +27,7 @@
         <label for="password" class="password-label">비밀번호</label>
         <div class="password-box">
           <input type="password" id="password" class="password-input" />
-          <button class="button-eye">
+          <button class="button-eye" type="button">
             <img src="images/eye-off.svg" alt="" />
           </button>
         </div>
@@ -60,5 +60,6 @@
         </div>
       </div>
     </div>
+    <script type="module" src="/src/signup.js"></script>
   </body>
 </html>
