+++ src/sign-style.css
@@ -77,6 +77,10 @@ input:focus {
   margin-top: 24px;
 }
 
+label.password-check {
+  margin-top: 24px;
+}
+
 .button-submit {
   border: none;
   display: flex;
