+++ src/utils.js
@@ -8,16 +8,22 @@ export const alertMessageBox = {
     noInput: '이메일을 입력해주세요.',
     inappropriateValue: '올바른 이메일 주소가 아닙니다.',
     loginFail: '이메일을 확인해주세요.',
+    takenEmail: '이미 사용 중인 이메일입니다.',
   },
   password: {
     noInput: '비밀번호를 입력해주세요.',
     loginFail: '비밀번호를 확인해주세요.',
+    simple: '비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요.',
+  },
+  passwordCheck: {
+    matchFail: '비밀번호가 일치하지 않아요.',
   },
 };
 
 export const emailInput = document.querySelector('.email-input');
 export const pswdInput = document.querySelector('.password-input');
-export const pswdEye = document.querySelector('.button-eye');
+export const pswdCheck = document.querySelector('#password-check');
+export const pswdEye = document.querySelectorAll('.button-eye');
 export const loginBtn = document.querySelector('.button-submit');
 
 // 에러 메세지가 발생해야 하는 위치, 에러 종류를 파라미터로 받고 메세지를 생성
@@ -32,6 +38,9 @@ export function makeErrorMessage(errorPlace, messageType) {
     case 'password':
       pswdInput.after(message);
       break;
+    case 'passwordCheck':
+      pswdCheck.after(message);
+      break;
   }
 }
 
@@ -44,6 +53,9 @@ export function eraseErrorMessage(errorPlace) {
       break;
     case 'password':
       placeToErase = pswdInput;
+      break;
+    case 'passwordCheck':
+      placeToErase = pswdCheck;
       break;
   }
   if (placeToErase.nextElementSibling.className === 'alert') {
